rest query for task and cf are work Start create activity NewTaskActivity

+++ app/src/main/java/com/nextstep/entity/PersonEntity.java
@@ -4,7 +4,7 @@ import java.util.HashSet;
 import java.util.Set;
 
 public class PersonEntity {
-    private int id;
+    private long id;
     private String name;
     private String email;
     private Set<TargetEntity> targets = new HashSet<>();
@@ -25,10 +25,10 @@ public class PersonEntity {
         this.email = email;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -63,10 +63,10 @@ public class PersonEntity {
 
     @Override
     public int hashCode() {
-        int result = id;
+        long result = id;
         result = 31 * result + (name != null ? name.hashCode() : 0);
         result = 31 * result + (email != null ? email.hashCode() : 0);
-        return result;
+        return (int) result;
     }
 
     public Set<TargetEntity> getTargets() {
